Improve performance for colorbar generation
The existing code would fill each line with values derived from
the colorbar table, but only 32-bits at a time.  Change the algorithm
to only do that for the first row, and then memcpy() the first row
to all subsequent rows.  This lets us take advantage of the optimized
code inside of memcpy (e.g. SSE vector operations, etc).

On my Macbook, this reduces the cost of colorbar generation for
1920x1080/30 FPS from 20% of the CPU down to 8%.