Move tracking of clocked out audio into libklbars
Because we want the sine wave to be aligned between calls, we have
logic which keeps track of where we left off in the previous call.

We originally prototyped this code in VLC, but it makes more sense
to push it down into libklbars where it can be shared by multiple
projects.