First pass at 10-bit refactoring
This patch allows for a 10-bit field to be used internally in
the library for generation of colorbars.  Before the internal
buffer was always 8-bit and we would colorspace convert to 10-bit
on finalize.  However this approach prevents us from building real
10-bit colorbars.

To supplement this change, we need to be able to colorspace
convert in both directions - to allow conversion of an 8-bit
internal buffer to a 10-bit final buffer and vice versa.

Because of this change, the API needed to be adjusted to allow the
caller to specify at initialization whether the internal buffer
should be 8 or 10 bit.